refactor(frontend): update all import paths and fix missing NotificationBell
