+++ frontend/src/App.jsx
@@ -3,9 +3,9 @@ import { BrowserRouter, Routes, Route, Navigate, Link } from "react-router-dom";
 import { AuthProvider } from "./context/AuthContext.jsx";
 import { NotificationProvider } from "./context/NotificationContext.jsx";
 import { ThemeProvider } from "./context/ThemeContext.jsx";
-import ProtectedRoute from "./components/ProtectedRoute.jsx";
-import Layout from "./components/Layout.jsx";
-import ErrorBoundary from "./components/ErrorBoundary.jsx";
+import ProtectedRoute from "./components/layout/ProtectedRoute.jsx";
+import Layout from "./components/layout/Layout.jsx";
+import ErrorBoundary from "./components/layout/ErrorBoundary.jsx";
 
 const Login = lazy(() => import("./pages/Login.jsx"));
 const Dashboard = lazy(() => import("./pages/Dashboard.jsx"));
