+++ frontend/src/components/crawl/CrawlDialsPanel.jsx
@@ -14,7 +14,7 @@
 
 import React, { useState } from "react";
 import { Settings2, ChevronDown, ChevronUp } from "lucide-react";
-import TestDials from "./TestDials.jsx";
+import TestDials from "../shared/TestDials.jsx";
 import { countActiveDials, loadSavedConfig } from "../utils/testDialsStorage.js";
 
 export default function CrawlDialsPanel({ value, onChange }) {
