+++ frontend/src/components/layout/Layout.jsx
@@ -2,10 +2,10 @@ import React from "react";
 import { Outlet } from "react-router-dom";
 import OnboardingTour from "./OnboardingTour.jsx";
 import useOnboarding from "../hooks/useOnboarding.js";
-import AIChat from "./AIChat.jsx";
+import AIChat from "../ai/AIChat.jsx";
 import CommandPalette from "./CommandPalette.jsx";
-import Sidebar from "./layout/Sidebar.jsx";
-import TopBar from "./layout/TopBar.jsx";
+import Sidebar from "./Sidebar.jsx";
+import TopBar from "./TopBar.jsx";
 
 export default function Layout() {
   const tour = useOnboarding();
