+++ frontend/src/pages/Applications.jsx
@@ -6,8 +6,8 @@ import {
 } from "lucide-react";
 import useProjectData, { invalidateProjectDataCache } from "../hooks/useProjectData";
 import { fmtRelativeDate } from "../utils/formatters";
-import PassRateBar from "../components/PassRateBar";
-import DeleteProjectModal from "../components/DeleteProjectModal.jsx";
+import PassRateBar from "../components/charts/PassRateBar";
+import DeleteProjectModal from "../components/shared/DeleteProjectModal.jsx";
 import { api } from "../api.js";
 import usePageTitle from "../hooks/usePageTitle.js";
